Added some db models, audit to employees and some service logic refactoring

+++ Human Capital Management/HCM.Data/Models/PositionSeniority.cs
@@ -6,7 +6,7 @@ namespace HCM.Data.Models
     public partial class PositionSeniority
     {
         public int? PositionId { get; set; }
-        public byte? SeniorityId { get; set; }
+        public int? SeniorityId { get; set; }
 
         public virtual Position? Position { get; set; }
         public virtual Seniority? Seniority { get; set; }
