+++ Human Capital Management/HCM.Models.ViewModels/Seniorities/SeniorityViewModel.cs
@@ -3,7 +3,7 @@
     public class SeniorityViewModel
     {
 
-        public byte? Id { get; set; }
+        public int? Id { get; set; }
 
         public string Name { get; set; }
